rearchitecture treatment methods from saveArticle.ts file
